requirements.txt
python-telegram-bot==20.7
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
cachetools==5.3.2
aiohttp==3.9.1
apscheduler==3.10.4
